V0.0.12.5

+++ app/src/main/java/com/example/dayone_calorietracker/Fragments/AddMealFragment.java
@@ -98,11 +98,11 @@ public class AddMealFragment extends Fragment {
                 });
             }else if("Confirm".equals(_Action)){
 
-                calories.setText(String.valueOf(getArguments().getInt("calories")));
-                protein.setText(String.valueOf(getArguments().getInt("protein")));
-                carbs.setText(String.valueOf(getArguments().getInt("carbs")));
-                fat.setText(String.valueOf(getArguments().getInt("fat")));
-                sugar.setText(String.valueOf(getArguments().getInt("sugar")));
+                calories.setText(String.valueOf(getArguments().getDouble("calories")));
+                protein.setText(String.valueOf(getArguments().getDouble("protein")));
+                carbs.setText(String.valueOf(getArguments().getDouble("carbs")));
+                fat.setText(String.valueOf(getArguments().getDouble("fat")));
+                sugar.setText(String.valueOf(getArguments().getDouble("sugar")));
             }
         }
